Autenticação e Validação com JWT

+++ tests/test_app.py
@@ -56,17 +56,11 @@ def test_create_user_email_already_exists(client, user):
     assert response.json() == {'detail': 'Email de usuário já existente'}
 
 
-def test_read_users(client):
-    response = client.get('/users/')
-
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'users': []}
-
-
-def test_read_users_with_users(client, user):
+def test_read_users(client, user, token):
 
     user_schema = UserPublic.model_validate(user).model_dump()
-    response = client.get('/users/')
+
+    response = client.get('/users/', headers={'Authorization': f'Bearer {token}'})
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {'users': [user_schema]}
@@ -90,13 +84,14 @@ def test_get_user(client, user):
     }
 
 
-def test_update_user(client, user):
+def test_update_user(client, user, token):
     client.post(
         '/users/', json={'username': 'arthur', 'email': 'arthur@example.com', 'password': '123'}
     )
 
     response = client.put(
         '/users/1',
+        headers={'Authorization': f'Bearer {token}'},
         json={
             'username': 'bob',
             'email': 'bob@example.com',
@@ -107,9 +102,10 @@ def test_update_user(client, user):
     assert response.status_code == HTTPStatus.OK
 
 
-def test_update_user_not_found(client):
+def test_update_user_not_found(client, token):
     response = client.put(
         '/users/999',
+        headers={'Authorization': f'Bearer {token}'},
         json={
             'username': 'bob',
             'email': 'bob@example.com',
@@ -117,13 +113,14 @@ def test_update_user_not_found(client):
         },
     )
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {'detail': 'Usuário não encontrado'}
+    assert response.status_code == HTTPStatus.FORBIDDEN
+    assert response.json() == {'detail': 'Sem permissões'}
 
 
-def test_update_integrity_error(client, user):
+def test_update_integrity_error(client, user, token):
     client.post(
         '/users',
+        headers={'Authorization': f'Bearer {token}'},
         json={
             'username': 'fausto',
             'email': 'fausto@example.com',
@@ -133,6 +130,7 @@ def test_update_integrity_error(client, user):
 
     response_update = client.put(
         f'/users/{user.id}',
+        headers={'Authorization': f'Bearer {token}'},
         json={
             'username': 'fausto',
             'email': 'bob@example.com',
@@ -144,15 +142,29 @@ def test_update_integrity_error(client, user):
     assert response_update.json() == {'detail': 'Username or Email já existem'}
 
 
-def test_delete_user(client, user):
-    response = client.delete('/users/1')
+def test_delete_user(client, user, token):
+    response = client.delete('/users/1', headers={'Authorization': f'Bearer {token}'})
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {'message': 'usuário deletado'}
 
 
-def test_delete_user_not_found(client):
-    response = client.delete('/users/999')
+def test_delete_user_not_found(client, token):
+    response = client.delete(
+        '/users/999',
+        headers={'Authorization': f'Bearer {token}'},
+    )
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {'detail': 'Usuário não encontrado'}
+    assert response.status_code == HTTPStatus.FORBIDDEN
+    assert response.json() == {'detail': 'Sem permissões'}
+
+
+def test_get_token(client, user):
+
+    response = client.post('/token', data={'username': user.email, 'password': user.clean_password})
+
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert token['token_type'] == 'bearer'
+    assert 'access_token' in token
